chore: managers empty state
